Added Column and Row Count Properties to Cell Grid Panel
I need to be able to change the number of rows and columns of the grid
at run time. This isn't much of an issue, but I also need a way to set
the cells into the grid. The easiest way I can think of this is to set
the grid as the panel of an items control (letting me bind the items
source to my cell collection). Once I do that, I will loose easy access
to the grid. I should be able to get access to it (somehow) in code, but
I'm going to try setting the column and row counts in a parent first and
see if that works.