Cell Grid Panel is now in use.
I started to use the Cell Grid Panel in the panel template of a List
Box, but setting the properties that change the number of columns and
rows weren't trickling down to the Cell Grid Panel. Template Binding
those properties to the List Box didn't work either, but now that I
think about it, I should have used a Relative Binding. I ended up
walking down the Visual Tree to find the Cell Grid Panel. Keeping the
code that updates the columns and rows wasn't need in the Cell Grid
Panel after that, but it was there and a normal grid wouldn't work. A
normal grid wouldn't work because there are other grids in the List Box,
and telling a Cell Grid Panel apart from a Grid is a lot easier than
telling a Grid apart from a Grid.