Added additional state to Cell Grid State, and it doesn't look like Formation Locked property is needed anymore in Cell Grid View Model
The Formation Locked property in the Cell Grid State was meant as a way
to look the cells from direct user interaction. That means, that it
would only be set to false during the initial generation (after all the
cells have loaded). We now have a Cell Grid State enum that holds the
different states that the cell grid is in. I might bring back the
Formation Locked property, but it doesn't look like it's needed anymore.