Adding a Bool To Convent Converter class.
The Bool to Content Converter class is something I came up with when I
was generalizing a Bool to Solid Color Brush Converter class I wrote for
a different project a few years ago.  For this project, the converter
will be used to change the content in the button that starts/stops the
auto progression in the game.