Working on Game Progression
I've created stub methods for the use of the commands that will control
the progression of the game once the User is finished setting the
initial cell generation. These methods have been tied into the commands
that will use them, and the controls that launch the commands have
already been bound to said commands.