Code commenting and minor changes
I renamed the method for finding the CellGridPanel in the VisualTree
from FindCellGrid to Find_CellGridPanel. The reason for this is because
there already is something else that is known as a cell grid, and I
didn't want the two to be confused. The under score in the name is just
there to help drive the point.