Removed unused code
I removed the Attacked Dependency Property, AcceptsManualInput because
it ended up not being used. I was going to use this property to control
when a User could change the living state of the cells, but I ended up
just making it so that a User can't click on the CellGridView and its
content.

I also removed a few binding that were going to be used for Grid
placement, but they were there because I tend to forget that I need to
set the Grid placement on the ListBoxItem, not the content item of the
ListBoxItem.